@@ -16,8 +16,10 @@
     "@react-navigation/native-stack": "^6.1.0",
     "react": "17.0.1",
     "react-native": "0.64.1",
+    "react-native-config": "^1.4.4",
     "react-native-safe-area-context": "^3.3.0",
-    "react-native-screens": "^3.5.0"
+    "react-native-screens": "^3.5.0",
+    "react-native-vector-icons": "^8.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
@@ -25,6 +27,7 @@
     "@react-native-community/eslint-config": "^2.0.0",
     "@types/jest": "^26.0.23",
     "@types/react-native": "^0.64.5",
+    "@types/react-native-vector-icons": "^6.4.8",
     "@types/react-test-renderer": "^16.9.2",
     "babel-jest": "^26.6.3",
     "eslint": "^7.14.0",
@@ -33,7 +36,7 @@
     "lint-staged": "^11.1.1",
     "metro-react-native-babel-preset": "^0.64.0",
     "react-test-renderer": "17.0.1",
-    "typescript": "^3.8.3"
+    "typescript": "^4.3.2"
   },
   "resolutions": {
     "@types/react": "^17"
