@@ -1,8 +1,9 @@
 import React from 'react';
-const { createNativeStackNavigator } = require('@react-navigation/native-stack');
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
 import AuthStack from './AuthStack';
 import TabNavigator from './TabNavigator';
+import { ExperimentalScreen } from '../screens';
 
 const Stack = createNativeStackNavigator();
 
@@ -10,6 +11,7 @@ const MainNavigator = () => (
   <Stack.Navigator initialRouteName="AuthStack" screenOptions={{ headerShown: false }}>
     <Stack.Screen name="AuthStack" component={AuthStack} />
     <Stack.Screen name="TabNavigator" component={TabNavigator} />
+    <Stack.Screen name="Experimental" component={ExperimentalScreen} />
   </Stack.Navigator>
 );
 
