@@ -1,15 +1,54 @@
 import React from 'react';
-const { createBottomTabNavigator } = require('@react-navigation/bottom-tabs');
+import MaterialIcon from 'react-native-vector-icons/MaterialIcons';
+import { RouteProp } from '@react-navigation/native';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 
-import { HistoryScreen, HomeScreen } from '../screens';
+import { HistoryScreen } from '../screens';
+import HomeStack from './HomeStack';
+
+import { colors } from '../utils/theme';
+
+type Route = RouteProp<Record<string, object | undefined>, string>;
 
 const Tab = createBottomTabNavigator();
 
+const getIconName = (routeName: string) => {
+  let iconName = '';
+  switch (routeName) {
+    case 'HomeTab':
+      iconName = 'home';
+      break;
+    case 'HistoryTab':
+      iconName = 'history';
+      break;
+    default:
+      iconName = 'help-outline';
+      break;
+  }
+
+  return iconName;
+};
+
+const navigatorScreenOptions = ({ route }: { route: Route }) => ({
+  tabBarIcon: ({ focused, color, size }: { focused: boolean; color: string; size: number }) => {
+    const iconName = getIconName(route.name);
+    const iconSize = focused ? size * 1.2 : size;
+    return <MaterialIcon name={iconName} size={iconSize} color={color} />;
+  },
+  tabBarAllowFontScaling: false,
+  tabBarActiveTintColor: colors.mainOrange,
+  tabBarInactiveTintColor: colors.lightBlue,
+  tabBarLabelStyle: {
+    fontSize: 12,
+  },
+  headerShown: false,
+});
+
 const TabNavigator = () => {
   return (
-    <Tab.Navigator>
-      <Tab.Screen name="HomeTab" component={HomeScreen} />
-      <Tab.Screen name="HistoryTab" component={HistoryScreen} />
+    <Tab.Navigator screenOptions={navigatorScreenOptions}>
+      <Tab.Screen name="HomeTab" component={HomeStack} options={{ title: 'Home' }} />
+      <Tab.Screen name="HistoryTab" component={HistoryScreen} options={{ title: 'History' }} />
     </Tab.Navigator>
   );
 };
