@@ -1,4 +1,5 @@
 export { default as AlertModal } from './AlertModal';
 export { default as DefaultButton } from './DefaultButton';
-export { default as Typography } from './Typography';
+export { default as Header } from './Header';
 export { default as Separator } from './Separator';
+export { default as Typography } from './Typography';
